+++ src/components/CoSign/Cosign.spec.jsx
@@ -7,7 +7,7 @@ import {ConfigContext, FormContext} from 'Context';
 import {BASE_URL, buildForm} from 'api-mocks';
 import mswServer from 'api-mocks/msw-server';
 import {mockSubmissionGet, mockSubmissionSummaryGet} from 'api-mocks/submissions';
-import {cosignRoutes} from 'routes';
+import cosignRoutes from 'routes/cosign';
 
 import Cosign from './Cosign';
