+++ src/routes/index.jsx
@@ -1,3 +1,1 @@
-export {createAppointmentRoutes, manageAppointmentRoutes} from './appointments';
-export {default as cosignRoutes} from './cosign';
-export {default as formRoutes} from './form';
+export {default} from './app';
